New Commit

+++ src/Main/Main.java
@@ -1,15 +1,15 @@
 package Main;
 
-import Operations.FoodOperations;
+import Services.FoodOperations;
 import Models.Food;
+import Services.FoodServices;
 
-import java.time.LocalDateTime;
 import java.util.Scanner;
 import java.util.List;
 
 public class Main {
     public static void main(String[] args) {
-        FoodOperations foodOps = new FoodOperations();
+        FoodServices foodOps = new FoodOperations();
         int flag = 0;
         try (Scanner sc = new Scanner(System.in)) {
             do {
@@ -19,6 +19,7 @@ public class Main {
                 System.out.println("4.Update Food");
                 System.out.println("5.Search Food");
                 System.out.println("6.Display Food");
+                System.out.println("7.Display in Json");
                 System.out.println("Enter your choice: ");
                 int choice = sc.nextInt();
                 switch (choice) {
@@ -31,7 +32,6 @@ public class Main {
                         addObj.setCategory(sc.nextLine());
                         System.out.println("Enter food price: ");
                         addObj.setPrice(sc.nextDouble());
-//                        addObj.setCreatedDate(LocalDateTime.now());
                         foodOps.addFood(addObj);
                         break;
                     case 2:
@@ -43,50 +43,63 @@ public class Main {
                     case 3:
                         System.out.println("Enter Food id to delete:");
                         int deleteId = sc.nextInt();
-                        foodOps.deleteData(deleteId);
+                        if (foodOps.isFoodIdExists(deleteId)) {
+                            foodOps.deleteData(deleteId);
+                        }
                         break;
                     case 4:
-                        Food updateObj = new Food();
                         System.out.println("Enter food id to update:");
-                        updateObj.setFoodId(sc.nextInt());
-                        System.out.println("Select what you want to update:");
-                        System.out.println("1.Food Name\t2.Category\t3.Price");
-                        System.out.println("Enter your choice to update:");
-                        int choiceUpdate = sc.nextInt();
-                        switch (choiceUpdate) {
-                            case 1:
-                                System.out.println("Enter food name to update: ");
-                                sc.nextLine();
-                                updateObj.setFoodName(sc.nextLine());
-                                foodOps.updateData(updateObj, choiceUpdate);
-                                break;
-                            case 2:
-                                System.out.println("Enter category to update: ");
-                                sc.nextLine();
-                                updateObj.setCategory(sc.nextLine());
-                                foodOps.updateData(updateObj, choiceUpdate);
-                                break;
-                            case 3:
-                                System.out.println("Enter price to update: ");
-                                updateObj.setPrice(sc.nextDouble());
-                                foodOps.updateData(updateObj, choiceUpdate);
-                                break;
-                            default:
-                                System.out.println("Invalid choice");
+                        int updateId = sc.nextInt();
+                        if (foodOps.isFoodIdExists(updateId)) {
+                            Food updateObj = new Food();
+                            updateObj.setFoodId(updateId);
+                            int foodId = updateObj.getFoodId();
+                            System.out.println("Select what you want to update:");
+                            System.out.println("1.Food Name\t2.Category\t3.Price");
+                            System.out.println("Enter your choice to update:");
+                            int choiceUpdate = sc.nextInt();
+                            switch (choiceUpdate) {
+                                case 1:
+                                    System.out.println("Enter food name to update: ");
+                                    sc.nextLine();
+                                    updateObj.setFoodName(sc.nextLine());
+                                    foodOps.updateData(updateObj, choiceUpdate);
+                                    break;
+                                case 2:
+                                    System.out.println("Enter category to update: ");
+                                    sc.nextLine();
+                                    updateObj.setCategory(sc.nextLine());
+                                    foodOps.updateData(updateObj, choiceUpdate);
+                                    break;
+                                case 3:
+                                    double oldPrice = foodOps.getOldPrice(foodId);
+                                    System.out.println("Enter price to update: ");
+                                    updateObj.setPrice(sc.nextDouble());
+                                    foodOps.updatePrice(updateObj, oldPrice);
+                                    break;
+                                default:
+                                    System.out.println("Invalid choice");
+                            }
                         }
+
                         break;
                     case 5:
                         sc.nextLine();
                         System.out.println("Enter food name to search: ");
                         String keyword = sc.nextLine();
                         List<Food> foodList = foodOps.searchFood(keyword);
+                        System.out.println("FoodId\tFoodName\tCategory\t\tPrice");
                         for (Food f : foodList) {
-                            System.out.println(f.getFoodId() + "\t" + f.getFoodName() + "\t" + f.getCategory() + "\t");
+                            System.out.println(f.getFoodId() + "\t" + f.getFoodName() + "\t" + f.getCategory() + "\t" + f.getPrice());
                         }
                         break;
                     case 6:
                         System.out.println("Displaying database");
                         foodOps.displayData();
+                        break;
+                    case 7:
+                        System.out.println("Displaying data in json");
+                        foodOps.displayWithJson();
                         break;
                     default:
                         System.out.println("Invalid choice");
